+++ wxc_sdk/cdr/__init__.py
@@ -541,7 +541,8 @@ class DetailedCDRApi(ApiChild, base=''):
     """
 
     def get_cdr_history(self, start_time: Union[str, datetime] = None, end_time: Union[datetime, str] = None,
-                        locations: list[str] = None, **params) -> Generator[CDR, None, None]:
+                        locations: list[str] = None, host:str = 'analytics.webexapis.com', stream: bool = False,
+                        **params) -> Generator[CDR, None, None]:
         """
         Provides Webex Calling Detailed Call History data for your organization.
 
@@ -563,10 +564,15 @@ class DetailedCDRApi(ApiChild, base=''):
         :param locations: Names of the location (as shown in Control Hub). Up to 10 comma-separated locations can be
             provided. Allows you to query reports by location.
         :type locations: list[str]
+        :param host: analytics host to access. Defaults to 'analytics.webexapis.com'.
+        :type host: str
+        :param stream: If true, collect data from cdr_stream, else from cdr_feed. Defaults to False (cdr_feed).
+        :type stream: bool
         :param params: additional arguments
         :return:
         """
-        url = 'https://analytics.webexapis.com/v1/cdr_feed'
+        endpoint = 'cdr_stream' if stream else 'cdr_feed'
+        url = f'https://{host}/v1/{endpoint}'
         if locations:
             params['locations'] = ','.join(locations)
         if not start_time: